♻️ 위치 정보 컴포넌트 분리
